Extract examples to another file
