+++ 03.event-emitter-buffer-streams/with-time.js
@@ -1,0 +1,3 @@
+import EventEmitter from './event-emitter.js'
+
+console.log(EventEmitter)
