+++ package.json
@@ -3,8 +3,10 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "m3t1": "nodemon 03.event-emitter-buffer-streams/event-emitter.js",
+    "m3t2": "nodemon 03.event-emitter-buffer-streams/with-time-examples.js",
+    "m3t1": "nodemon 03.event-emitter-buffer-streams/event-emitter-examples.js",
     "m2": "nodemon 02.standard-library/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
